feat:create PagedTransactionMapper.kt

+++ wallet_domain/src/commonMain/kotlin/net/thechance/mena/wallet/domain/entity/PagedTransactionResponse.kt
@@ -1,10 +1,10 @@
 package net.thechance.mena.wallet.domain.entity
 
 data class PagedTransactionResponse(
-    val totalElements: Int,
-    val page: Int,
-    val pageSize: Int,
-    val totalPages: Int,
+    val totalElements: Long,
+    val page: Long,
+    val pageSize: Long,
+    val totalPages: Long,
     val transactions: List<Transaction>
 
 )
